Extract printInterface and printCallExpression (#9715)
